+++ lib/oauth2.dart
@@ -5,6 +5,7 @@
 library oauth2;
 
 export 'src/authorization_code_grant.dart';
+export 'src/resource_owner_password_grant.dart';
 export 'src/client.dart';
 export 'src/credentials.dart';
 export 'src/authorization_exception.dart';